Add distance check for double-tap
Previously only the time between taps was considered when trying to
detect double-taps. As a result the gesture handle could be tapped on
both ends and the last tap would act as a double-tap to cycle in that
direction.

Instead we now check if the double-tap is within a certain distance of
the original tap, behaving more consistently with what a double-tap is
commonly expected to be

Additionally a small bug was fixed where it was possible that the tap
after a double-tap would be recognized as a double-tap, since the
double-tap itself was considered the start of a new tap sequence.

There's also been a minor refactor of the `set_sleep` methods used to
control display power, since they are completely unrelated to sleep.