Add support for double-tap display wakeup
This adds support for waking up the display of a non-suspended system by
double-tapping the screen, making it a little easier to recover when the
display timeout is reached while using the phone.

Since this is generally a less precise action, this patch also increases
the maximum distance for double-taps to about double of what it was
previously.

Closes #156.