Fix DPMS handling regressions
This patch fixes two separate regressions. The first one was introduced
in 3d04f9ee, where the `sleeping` field was renamed to `display_on`.
Since the field's value was inverted, it was necessary to switch all
conditionals accessing this field, which was forgotten for the one that
toggles the accelerometer timer on DPMS on/off.

Then 64279b21 made use of the same conditional to simulate activity
whenever the display is turned on, however it accidentally checked for
the display to be off. As a result turning the display off would
automatically cause activity, waking the display back up.