Fix screencopy ignoring orientation transactions
This fixes an issue where the SHM screencopy implementation would
directly use the output's current rotation, rather than using the
transactional orientation, which lead to single frame glitches when
recording orientation changes.

To fix this issue and make it more difficult for similar issues to exist
in the future, the `Windows::output` method was removed entirely, which
means only the windows module itself has access to non-transactional
output transforms now.

Closes #167.