Fix input region calculations
This fixes an issue in calculating which surface should receive input,
which was caused by layer shell surfaces always being the only surfaces
considered for input if the touch position was within their window
geometry.

To work properly, layer shell window input is now only consumed by the
layer shell window if one of the surfaces actually receives the input
taking surface an input regions into account.

This might be the cause of #157.