Disable render time predictor
This disables the frame pacer which is responsible for scheduling
rendering as close to VBlank as possible to optimize input latency,
since it was having the opposite effect by consistently missing the
VBlank interval due to failing to account for async rendering time.

This currently is just disabled temporarily until a proper solution is
implemented which keeps accurate track of the time it took to render the
frame.