Fix frame scheduling
This reimplements the frame scheduling algorithm, properly taking into
account the asynchronous processing time taken by the GPU by waiting for
the primary element's fence before completing the measurement of the
current frame.

Since this improves accuracy of the frame scheduler, the prediction
padding has also been changed from a fixed 8ms, to a flexible 150% of
the original prediction.

With an average render time of 8ms in PP tests, this causes the final
prediction to come out at ~12ms, which is half of the previous padding.
When comparing 150% to 125%, a significant increase in missed VBlanks
was observed, while a comparison to 200% showed no improvement.

Closes #178.