Add support for multi-key bindings
This allows binding commands to multi-key bindings by supporting the
concatenation of keysyms with `+` in the `bind-key` IPC message.

While this will not trigger individual key bindings while multiple keys
are being pressed, it will trigger `press` and `release` bindings for
the initial key.

Since the repeat timer is cancelled whenever a new keyboard event is
received, this will *not* trigger `repeat` bindings except for the
initial implicit `press` trigger.

The default initrc is updated to use this new mechanism to allow taking
screenshots when pressing `power + volume down`.

Closes #179.