Forward subprocess output
This changes the way output for daemons is handled by forwarding their
output stream to catacomb's stdout/stderr rather than dropping all
subprocess output. This should help with debugging issues with the
config and bindings.

This also removes a duplicate implementation of the daemon function.

Additionally the signals sent to kill `sleep` in the default
configuration was changed from `SIGKILL` to `SIGINT`, since the latter
doesn't generate a message from the shell that a child process was
killed.